# target/classes/appSwagger.yaml
openapi: 3.0.0
info:
  title: Title
  description: Title
  version: 1.0.0
servers:
  - url: 'https'
paths:
  /topsecret:
    post:
      summary: obtains the location and the emited message of the ship
      operationId: obtainDataShip
      requestBody:
        description: satellites
        required: true
        content:
          aplication/json:
            schema:
              type: object
              title: dataShipRequest
              properties:
                satellites:
                  type: array
                  items:
                    $ref: '#/components/schemas/SatelliteRequest'
      responses:
        200:
          description: Ok
          content:
            aplication/json:
              schema:
                type: object
                title: dataShipResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Position'
                  message:
                    type: string
        404:
          description: The message or the possition couldn't be determinated
  /topsecret_split:
    get:
      summary: obtains the location and the emited message of the ship
      operationId: obtainSplitDataShip
      responses:
        200:
          description: Ok
          content:
            aplication/json:
              schema:
                type: object
                title: dataShipResponse
                properties:
                  position:
                    $ref: '#/components/schemas/Position'
                  message:
                    type: string
        404:
          description: The message or the possition couldn't be determinated
  /topsecret_split/{satellite_name}:
    post:
      summary: receives different parts of the message
      operationId: receiveDataShip
      parameters:
        - name: satellite_name
          in: path
          description: satellite to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          aplication/json:
            schema:
              type: object
              title: satelliteSplitRequest
              properties:
                distance:
                  type: number
                  format: float
                message:
                  type: array
                  items:
                    type: string
      responses:
        200:
          description: Ok
          content:
            aplication/json:
              schema:
                type: object
                title: satelliteSplitResponse
                properties:
                  position:
                    $ref: '#/components/schemas/SatelliteRequest'
                  message:
                    type: string

components:
  schemas:
    SatelliteRequest:
      type: object
      properties:
        name:
          type: string
        distance:
          type: number
          format: float
        message:
          type: array
          items:
            type: string
    Position:
      type: object
      properties:
        x:
          type: number
          format: float
        y:
          type: number
          format: float





  